social_platform/requirements.txt
django
djangorestframework
django-cors-headers
django-channels[dapne]
channels-redis
pillow
psycopg2-binary
EOT